requirements.txt
Flask==3.0.3
opencv-python-headless==4.10.0.84
Pillow==10.4.0
ultralytics==8.2.73
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
supervision==0.1.5
werkzeug==2.3.4
